first cut add more tests for store
